start colocating zeromq native stuff together
